# Conflicts:
#   portfolio/src/main/java/nz/ac/canterbury/seng302/portfolio/controller/AccountController.java
#   portfolio/src/main/resources/static/javascript/editAccount.js
#   portfolio/src/main/resources/templates/account.html
#   portfolio/src/main/resources/templates/editAccount.html

+++ portfolio/src/main/java/nz/ac/canterbury/seng302/portfolio/controller/DashboardController.java
@@ -1,38 +1,51 @@
 package nz.ac.canterbury.seng302.portfolio.controller;
 
 import nz.ac.canterbury.seng302.portfolio.model.Project;
-import nz.ac.canterbury.seng302.portfolio.model.Sprint;
 import nz.ac.canterbury.seng302.portfolio.service.DashboardService;
 import nz.ac.canterbury.seng302.portfolio.service.SprintService;
+import nz.ac.canterbury.seng302.portfolio.service.UserAccountClientService;
+import nz.ac.canterbury.seng302.shared.identityprovider.AuthState;
+import nz.ac.canterbury.seng302.shared.identityprovider.ClaimDTO;
+import nz.ac.canterbury.seng302.shared.identityprovider.UserRole;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.time.LocalDate;
-
+import java.util.Arrays;
 import java.util.List;
 
-// TODO:
+
 @Controller
 public class DashboardController {
     @Autowired private DashboardService dashboardService;
+    @Autowired private UserAccountClientService userAccountClientService;
     @Autowired private SprintService sprintService;
 
 
     /**
-     * Adds project list to model and opens dashboard.html
+     * Adds the project list to the Dashboard.
+     * Gets the list of the current user's roles
+     * Opens dashboard.html
      * @param model
-     * @return
+     * @param principal - current user.
+     * @return name of the dashboard html file.
      */
     @GetMapping("/dashboard")
-    public String showProjectList(Model model) {
+    public String showProjectList( @AuthenticationPrincipal AuthState principal,
+                                   Model model) {
         List<Project> listProjects = null;
         try {
             listProjects = dashboardService.getAllProjects();
             model.addAttribute("listProjects", listProjects);
+            // Add the list of the current users roles to model so they can be used in dashboard.html with thymeleaf.
+            model.addAttribute("roles", userAccountClientService.getUserRole(principal));
             return "dashboard";
         } catch (Exception e) {
             model.addAttribute("exception", e);
@@ -52,9 +65,15 @@ public class DashboardController {
      * @return
      */
     @GetMapping("/dashboard/newProject")
-    public String showNewForm(Model model) {
+    public String showNewForm(Model model, @AuthenticationPrincipal AuthState principal) {
+        List<String> userRoles = Arrays.asList(principal.getClaimsList().stream().filter(claim -> claim.getType().equals("role")).findFirst().map(ClaimDTO::getValue).orElse("NOT FOUND").split(","));
+        if (!(userRoles.contains(UserRole.TEACHER.name()) || userRoles.contains(UserRole.COURSE_ADMINISTRATOR.name()))) {
+            return "redirect:/dashboard";
+        }
+        dashboardService.setPreviousFeature("Create");
         model.addAttribute("project", new Project());
         model.addAttribute("pageTitle", "Add New Project");
+        model.addAttribute("submissionName", "Create");
         return "projectForm";
     }
 
@@ -64,9 +83,19 @@ public class DashboardController {
      * @return
      */
     @PostMapping("/dashboard/saveProject")
-    public String saveProject(Project project, Model model) {
+    public String saveProject(Project project, Model model, RedirectAttributes ra, @AuthenticationPrincipal AuthState principal) {
+        List<String> userRoles = Arrays.asList(principal.getClaimsList().stream().filter(claim -> claim.getType().equals("role")).findFirst().map(ClaimDTO::getValue).orElse("NOT FOUND").split(","));
+        if (!(userRoles.contains(UserRole.TEACHER.name()) || userRoles.contains(UserRole.COURSE_ADMINISTRATOR.name()))) {
+            return "redirect:/dashboard";
+        }
         try {
+            String msgString;
             dashboardService.saveProject(project);
+            msgString = dashboardService.createSuccessMessage(project);
+            if (msgString == null) {
+                throw new Exception("Error generating success message");
+            }
+            ra.addFlashAttribute("messageSuccess", msgString);
             return "redirect:/dashboard";
         } catch (Exception e) {
             model.addAttribute("exception", e);
@@ -87,11 +116,25 @@ public class DashboardController {
      * @return
      */
     @GetMapping("/dashboard/editProject/{projectId}")
-    public String showEditForm(@PathVariable(value = "projectId") int projectId, Model model) {
-        Project project  = dashboardService.getProject(projectId);
-        model.addAttribute("project", project);
-        model.addAttribute("pageTitle", "Edit Project (Name: " + projectId + ")");
-        return "projectForm";
+    public String showEditForm(
+        @PathVariable(value = "projectId") int projectId, 
+        Model model, RedirectAttributes ra,
+        @AuthenticationPrincipal AuthState principal) {
+        List<String> userRoles = Arrays.asList(principal.getClaimsList().stream().filter(claim -> claim.getType().equals("role")).findFirst().map(ClaimDTO::getValue).orElse("NOT FOUND").split(","));
+        if (!(userRoles.contains(UserRole.TEACHER.name()) || userRoles.contains(UserRole.COURSE_ADMINISTRATOR.name()))) {
+            return "redirect:/dashboard";
+        }
+        try {
+            dashboardService.setPreviousFeature("Edit");
+            Project project  = dashboardService.getProject(projectId);
+            model.addAttribute("project", project);
+            model.addAttribute("pageTitle", "Edit Project: " + project.getProjectName());
+            model.addAttribute("submissionName", "Save");
+            return "projectForm";
+        } catch (NullPointerException e) {
+            ra.addFlashAttribute("messageDanger", "No Project Found");
+            return "redirect:/dashboard";
+        }
     }
 
     /**
@@ -101,7 +144,14 @@ public class DashboardController {
      * @throws Exception If project is not found in the database
      */
     @GetMapping("/dashboard/deleteProject/{projectId}")
-    public String deleteProject(@PathVariable("projectId") int projectId, Model model) {
+    public String deleteProject(
+        @PathVariable("projectId") int projectId,
+        Model model,
+        @AuthenticationPrincipal AuthState principal) {
+        List<String> userRoles = Arrays.asList(principal.getClaimsList().stream().filter(claim -> claim.getType().equals("role")).findFirst().map(ClaimDTO::getValue).orElse("NOT FOUND").split(","));
+        if (!(userRoles.contains(UserRole.TEACHER.name()) || userRoles.contains(UserRole.COURSE_ADMINISTRATOR.name()))) {
+            return "redirect:/dashboard";
+        }
         try {
             sprintService.deleteAllSprints(projectId);
             dashboardService.deleteProject(projectId);
@@ -116,8 +166,5 @@ public class DashboardController {
             model.addAttribute("status", "re-run Project");
             return "error";
         }
-
-
     }
-
 }
